+++ packages/plugin-model/src/utils/getUseModelContent.ts
@@ -33,26 +33,47 @@ export function useModel<T extends keyof Model<T>, U>(
   const stateRef = useRef<any>(state);
   stateRef.current = state;
 
+  const isMount = useRef(false);
+  useEffect(() => {
+    isMount.current = true;
+    return () => {
+      isMount.current = false;
+    }
+  }, [])
+
   useEffect(() => {
     const handler = (e: any) => {
-      if(updater && updaterRef.current){
-        const currentState = updaterRef.current(e);
-        const previousState = stateRef.current
-        if(!isEqual(currentState, previousState)){
-          setState(currentState);
-        }
+      if(!isMount.current) {
+        // 如果 handler 执行过程中，组件被卸载了，则强制更新全局 data
+        setTimeout(() => {
+          dispatcher.data![namespace] = e;
+          dispatcher.update(namespace);
+        });
       } else {
-        setState(e);
+        if(updater && updaterRef.current){
+          const currentState = updaterRef.current(e);
+          const previousState = stateRef.current
+          if(!isEqual(currentState, previousState)){
+            setState(currentState);
+          }
+        } else {
+          setState(e);
+        }
       }
     }
     try {
       dispatcher.callbacks![namespace]!.add(handler);
+      dispatcher.update(namespace);
     } catch (e) {
       dispatcher.callbacks![namespace] = new Set();
       dispatcher.callbacks![namespace]!.add(handler);
+      dispatcher.update(namespace);
     }
     return () => {
-      dispatcher.callbacks![namespace]!.delete(handler);
+      // 保证组件卸载前，还能最后一次触发 handler
+      setTimeout(() => {
+        dispatcher.callbacks![namespace]!.delete(handler);
+      })
     }
   }, [namespace]);