+++ packages/plugin-qiankun/.fatherrc.ts
@@ -1,5 +1,6 @@
 export default {
   runtimeHelpers: true,
+  disableTypeCheck: false,
   browserFiles: [
     'src/master/runtimePlugin.ts',
     'src/slave/lifecycles.ts',